fix: cast iface ID to long unsigned int for proper logging in ciot_iface_send_rsp

+++ src/core/ciot_iface.c
@@ -35,7 +35,7 @@ ciot_err_t ciot_iface_send_rsp(ciot_iface_t *self, ciot_msg_t *rsp)
 
     if (self->req_status.state != CIOT_IFACE_REQ_STATE_IDLE)
     {
-        CIOT_LOGE(TAG, "Iface %s (%lu) is busy", ciot_iface_to_str(self), self->info.id);
+        CIOT_LOGE(TAG, "Iface %s (%lu) is busy", ciot_iface_to_str(self), (long unsigned int)self->info.id);
         return CIOT_ERR_BUSY;
     }
     else
